perf(app): improve animation
