+++ packages/app/src/game/components/field-cell/field-cell.tsx
@@ -1,27 +1,18 @@
-import { use, useImperativeHandle } from 'react';
+import { use } from 'react';
 import { Pressable } from 'react-native';
-import Reanimated, {
-    interpolate,
-    interpolateColor,
-    useAnimatedStyle,
-    useDerivedValue,
-    useSharedValue,
-    withTiming
-} from 'react-native-reanimated';
+import Reanimated, { interpolateColor, useAnimatedStyle, useDerivedValue, withTiming } from 'react-native-reanimated';
 
 import { type OnEventFn, cs } from '@rnw-community/shared';
 
 import { animationDurationConstant } from '../../../@generic/constants/animation.constant';
 import { ThemeContext } from '../../../@generic/context/theme.context';
 import { GameContext } from '../../context/game.context';
-import { CellFontSizeConstant } from '../constants/dimensions.contant';
 
 import { FieldCellSelectors as selectors } from './field-cell.selectors';
 import { FieldCellStyles as styles } from './field-cell.styles';
 
 import type { BlackTheme } from '../../../@generic/styles/theme';
 import type { CellInterface } from '@suuudokuuu/generator';
-import type { Ref } from 'react';
 
 const ReanimatedPressable = Reanimated.createAnimatedComponent(Pressable);
 
@@ -106,13 +97,6 @@ const getCandidateTextStyles = (theme: typeof BlackTheme, candidate: number) => 
 };
 
 const animationConfig = { duration: animationDurationConstant };
-const textAnimationConfig = { duration: 6 * animationDurationConstant };
-
-const FONT_SIZE_MULTIPLIER = 1.5;
-
-export interface FieldCellRef {
-    triggerAnimation: () => void;
-}
 
 interface Props {
     readonly cell: CellInterface;
@@ -122,55 +106,26 @@ interface Props {
     readonly isHighlighted: boolean;
     readonly isWrong: boolean;
     readonly hasCandidates: boolean;
-    readonly ref: Ref<FieldCellRef>;
+    readonly hasAnimation: boolean;
+    readonly textAnimatedStyle: ReturnType<typeof useAnimatedStyle>;
 }
 
 // TODO: Components is overloaded - split into smaller components
 // eslint-disable-next-line max-statements
 export const FieldCell = (props: Props) => {
-    const { cell, onSelect, isActive, isActiveValue, isHighlighted, isWrong, hasCandidates, ref } = props;
+    const { cell, onSelect, isActive, isActiveValue, isHighlighted, isWrong, hasCandidates, hasAnimation, textAnimatedStyle } = props;
 
     const { sudoku } = use(GameContext);
     const { theme } = use(ThemeContext);
 
     const isEmpty = sudoku.isBlankCell(cell);
     const cellBackgroundColor = getCellBgColor(theme, isActiveValue, isHighlighted, isWrong, isEmpty);
-    const cellTextColor = getCellTextColor(theme, isActive, isEmpty, isHighlighted, isActiveValue);
-    const text = getText(isActive, isEmpty, hasCandidates, cell);
     const candidates = isEmpty && hasCandidates ? sudoku.getCellCandidates(cell) : [];
 
     const animation = useDerivedValue(() => withTiming(isActive ? 1 : 0, animationConfig));
-    const textAnimation = useSharedValue(0);
-
-    useImperativeHandle(
-        ref,
-        () => ({
-            triggerAnimation: () => {
-                textAnimation.value = withTiming(1, textAnimationConfig, finished => {
-                    if (finished === true) {
-                        textAnimation.value = 0;
-                    }
-                });
-            }
-        }),
-        [textAnimation]
-    );
 
     const cellAnimatedStyles = useAnimatedStyle(() => ({
         backgroundColor: interpolateColor(animation.value, [0, 1], [cellBackgroundColor, theme.colors.cell.active])
-    }));
-    const textAnimatedStyles = useAnimatedStyle(() => ({
-        color: interpolateColor(
-            textAnimation.value,
-            [0, 0.5, 1],
-            [theme.colors.black, theme.colors.cell.highlightedText, theme.colors.black]
-        ),
-        fontSize: interpolate(
-            textAnimation.value,
-            [0, 0.5, 1],
-            [CellFontSizeConstant, CellFontSizeConstant * FONT_SIZE_MULTIPLIER, CellFontSizeConstant]
-        ),
-        transform: [{ rotate: `${interpolate(textAnimation.value, [0, 1], [0, 360])}deg` }]
     }));
 
     const handlePress = () => {
@@ -187,7 +142,12 @@ export const FieldCell = (props: Props) => {
         { backgroundColor: cellBackgroundColor },
         cellAnimatedStyles
     ];
-    const textStyles = [styles.textRegular, { color: cellTextColor }, cs(isActive, styles.textActive), textAnimatedStyles];
+    const textStyles = [
+        styles.textRegular,
+        { color: getCellTextColor(theme, isActive, isEmpty, isHighlighted, isActiveValue) },
+        cs(isActive, styles.textActive),
+        cs(hasAnimation, textAnimatedStyle)
+    ];
 
     return (
         <ReanimatedPressable onPress={handlePress} style={cellStyles} testID={getCellSelector(props)}>
@@ -197,7 +157,7 @@ export const FieldCell = (props: Props) => {
                 </Reanimated.Text>
             ))}
 
-            <Reanimated.Text style={textStyles}>{text}</Reanimated.Text>
+            <Reanimated.Text style={textStyles}>{getText(isActive, isEmpty, hasCandidates, cell)}</Reanimated.Text>
         </ReanimatedPressable>
     );
 };
