+++ packages/app/src/game/components/field/field.tsx
@@ -1,11 +1,15 @@
 import { isEmptyScoredCells } from '@suuudokuuu/generator';
-import { use, useImperativeHandle, useRef } from 'react';
+import { use, useImperativeHandle, useState } from 'react';
 import { View } from 'react-native';
+import { interpolate, interpolateColor, useAnimatedStyle, useSharedValue, withTiming } from 'react-native-reanimated';
 
+import { animationDurationConstant } from '../../../@generic/constants/animation.constant';
+import { ThemeContext } from '../../../@generic/context/theme.context';
 import { useAppSelector } from '../../../@generic/hooks/use-app-selector.hook';
 import { GameContext } from '../../context/game.context';
 import { gameHasCandidatesSelector } from '../../store/game.selectors';
-import { FieldCell, type FieldCellRef } from '../field-cell/field-cell';
+import { CellFontSizeConstant } from '../constants/dimensions.contant';
+import { FieldCell } from '../field-cell/field-cell';
 
 import { FieldStyles as styles } from './field.styles';
 
@@ -18,6 +22,8 @@ const getCellKey = (cell: CellInterface) => `${cell.y}-${cell.x}`;
 export interface FieldRef {
     triggerCellAnimations: (scoredCells: ScoredCellsInterface) => void;
 }
+const textAnimationConfig = { duration: 6 * animationDurationConstant };
+const FONT_SIZE_MULTIPLIER = 1.5;
 
 interface Props {
     readonly selectedCell?: CellInterface;
@@ -27,31 +33,54 @@ interface Props {
 
 export const Field = ({ selectedCell, onSelect, ref }: Props) => {
     const { sudoku } = use(GameContext);
+    const { theme } = use(ThemeContext);
     const hasCandidates = useAppSelector(gameHasCandidatesSelector);
 
-    const cellRefs = useRef<Record<string, FieldCellRef | null>>({});
+    const [animatedCells, setAnimatedCells] = useState(new Set<string>());
+
+    const textAnimation = useSharedValue(0);
+    const textAnimatedStyles = useAnimatedStyle(() => ({
+        color: interpolateColor(
+            textAnimation.value,
+            [0, 0.5, 1],
+            [theme.colors.black, theme.colors.cell.highlightedText, theme.colors.black]
+        ),
+        fontSize: interpolate(
+            textAnimation.value,
+            [0, 0.5, 1],
+            [CellFontSizeConstant, CellFontSizeConstant * FONT_SIZE_MULTIPLIER, CellFontSizeConstant]
+        ),
+        transform: [{ rotate: `${interpolate(textAnimation.value, [0, 1], [0, 360])}deg` }]
+    }));
 
     useImperativeHandle(
         ref,
         () => ({
             triggerCellAnimations: (scoredCells: ScoredCellsInterface) => {
+                const newAnimatedCells = new Set<string>();
                 if (!isEmptyScoredCells(scoredCells)) {
+                    textAnimation.value = 0;
+                    textAnimation.value = withTiming(1, textAnimationConfig, finished => {
+                        // TODO: Improve me
+                        if (finished === true) {
+                            textAnimation.value = 0;
+                        }
+                    });
+
                     sudoku.Field.forEach(row => {
                         row.forEach(cell => {
                             if (sudoku.isScoredCell(cell, scoredCells)) {
-                                cellRefs.current[getCellKey(cell)]?.triggerAnimation();
+                                newAnimatedCells.add(getCellKey(cell));
                             }
                         });
                     });
+
+                    setAnimatedCells(newAnimatedCells);
                 }
             }
         }),
-        []
+        [sudoku, textAnimation]
     );
-
-    const handleCellRef = (cell: CellInterface) => (cellRef: FieldCellRef | null) => {
-        cellRefs.current[getCellKey(cell)] = cellRef;
-    };
 
     return (
         <View style={styles.wrapper}>
@@ -60,6 +89,7 @@ export const Field = ({ selectedCell, onSelect, ref }: Props) => {
                     {row.map(cell => (
                         <FieldCell
                             cell={cell}
+                            hasAnimation={animatedCells.has(getCellKey(cell))}
                             hasCandidates={hasCandidates}
                             isActive={sudoku.isSameCell(cell, selectedCell)}
                             isActiveValue={sudoku.isSameCellValue(cell, selectedCell)}
@@ -67,7 +97,7 @@ export const Field = ({ selectedCell, onSelect, ref }: Props) => {
                             isWrong={sudoku.isCellWrong(cell, selectedCell)}
                             key={`cell-${cell.y}-${cell.x}`}
                             onSelect={onSelect}
-                            ref={handleCellRef(cell)}
+                            textAnimatedStyle={textAnimatedStyles}
                         />
                     ))}
                 </View>
